# coze-huiyan-muzong (1)/openapi.yaml
openapi: 3.0.0
info:
  title: 慧眼牧踪数据服务
  description: 慧眼牧踪智能养殖系统的数据接口，提供羊只管理、定位追踪、健康监测、预警管理等功能
  version: 1.0.0
servers:
  - url: https://你的公网域名
    description: Railway部署后的地址
security:
  - BearerAuth: []
paths:
...
  /api/sheep:
    get:
      operationId: getSheepList
      summary: 查询羊只列表
      description: 获取羊只分页列表，支持关键词、品种、健康状态筛选
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: pageSize
          in: query
          schema:
            type: integer
            default: 20
        - name: keyword
          in: query
          schema:
            type: string
          description: 搜索关键词（耳标号/品种）
        - name: breed
          in: query
          schema:
            type: string
          description: 品种筛选
        - name: health_status
          in: query
          schema:
            type: string
            enum: [健康, 观察中, 异常]
          description: 健康状态筛选
      responses:
        "200":
          description: 成功
    post:
      operationId: createSheep
      summary: 新增羊只
      description: 添加一只新的羊只记录
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tag_number, breed]
              properties:
                tag_number:
                  type: string
                  description: 耳标号
                breed:
                  type: string
                  description: 品种
                age:
                  type: integer
                  description: 年龄
                weight:
                  type: number
                  description: 体重(kg)
                gender:
                  type: string
                  enum: [雄, 雌]
                health_status:
                  type: string
                  enum: [健康, 观察中, 异常]
                herd_id:
                  type: integer
                  description: 所属羊群ID
                birth_date:
                  type: string
                  description: 出生日期
      responses:
        "200":
          description: 成功
...
  /api/trace/query:
    get:
      operationId: traceQuery
      summary: 溯源查询
      description: 通过溯源码查询产品全链路信息（养殖→检疫→加工→销售）
      parameters:
        - name: code
          in: query
          required: true
          schema:
            type: string
          description: 溯源码
      responses:
        "200":
          description: 成功
  /api/trace/sheep-code/{id}:
    get:
      operationId: getSheepTraceCode
      summary: 获取羊只溯源码
      description: 根据羊只ID获取其溯源码
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: 成功

  # ========== 系统设置 ==========
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: JWT令牌或Coze专用API Key